// src/api/models/tag.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * Aris-blog
 * OpenAPI spec version: 1.0
 */

export interface Tag {
  /** Creation timestamp */
  createdAt?: string;
  /** Tag description */
  description?: string;
  /**
   * Number of likes
   * @minimum 0
   */
  likes?: number;
  /** Tag name */
  name: string;
  /** Tag slug */
  slug: string;
  /**
   * Tag ID
   * @minimum 0
   */
  tagID: number;
  /** Update timestamp */
  updatedAt?: string;
}